Add files via upload
Objective
The objective of this assignment is to extract textual data articles from the given URL and perform text analysis to compute variables that are explained below. 

Data Extraction
Input.xlsx
For each of the articles, given in the input.xlsx file, extract the article text and save the extracted article in a text file with URL_ID as its file name.
Data Analysis
For each of the extracted texts from the article, perform textual analysis and compute variables, given in the output structure Excel file. You need to save the output in the exact order as given in the output structure file, “Output Data Structure.xlsx”
Variables
Definition of each of the variables given in the “Text Analysis.docx” file.
Look for these variables in the analysis document (Text Analysis.docx):
1.POSITIVE SCORE
2.NEGATIVE SCORE
3.POLARITY SCORE
4.SUBJECTIVITY SCORE
5.AVG SENTENCE LENGTH
6.PERCENTAGE OF COMPLEX WORDS
7.FOG INDEX
8.AVG NUMBER OF WORDS PER SENTENCE
9.COMPLEX WORD COUNT
10.WORD COUNT
11.SYLLABLE PER WORD
12.PERSONAL PRONOUNS
13.AVG WORD LENGTH